requirements.txt
ipdb==0.11
psycopg2==2.8.6
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
socketclusterclient==1.2.2
Jinja2==2.10